Changed the date of drawing with franka project
I wanted the projects to show up in order of importance and difficulty,
at least a little bit. So I switched the drawing with franka project to
have a later publishing date. After looking at the google analytics,
people click on the projects from top down, and the project that's clicked
the most is the Realistic Physics Sim.